requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
google-generativeai
pinecone-client
faiss-cpu
PyPDF2
numpy
sentence-transformers
python-dotenv
pytest
pytest-asyncio
httpx
redis
langchain
langchain-google-genai
langgraph
reportlab
aiofiles
python-jose[cryptography]
passlib[bcrypt]
